requirements.txt
streamlit
pandas
openpyxl
plotly
fpdf
Pillow
kaleido
